backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-google-genai==0.0.9
google-generativeai==0.3.1
requests==2.31.0
markdownify==0.11.6
google-search-results==2.4.2
chromadb>=0.4.24
python-multipart==0.0.9
pandas==2.2.0
sentence-transformers==2.2.2